requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
werkzeug
psycopg2-binary
flask-wtf
apscheduler
flask-mail
gunicorn
python-slugify